requirements.txt
Flask
yfinance
numpy
pandas
scikit-learn
matplotlib
tensorflow
requests
google-api-python-client
praw
yfinance
BeautifulSoup
openai
selenium
webdriver-manager
fake_useragent
torch
seaborn
python-dotenv
joblib
torch 
torchvision